Dice enhancements (#1331)
* first pass at dice upgrade/fix

* fix annotation

* shorted overlap_mask arg name

* move dice op to make it a proper lossOp

* fix per-instance logging for dice

* prevent missing objects (which are correctly identified as missing) from dragging down dice scores

* unused variable

* let ValWithError function more as a regular number

* fix single-channel per-instance dice logging

* fix std and valwitherror

* fix training whitespace

* fix issue with ValWithError + earlystopping

* make ValWithError work with restorewizard

* fix plotting of ValWithError

* modify dice casting for mixed precision

* fix some other cast operations

* fix torch cast warning message

* improve dice compatibility with mixed precision

* don't bother upcasting in trace b/c backend is based on mixed precision mode

* fix unused import

* pre-cast weights for marginal speed improvement

+++ fastestimator/backend/_binary_crossentropy.py
@@ -84,14 +84,14 @@ def binary_crossentropy(y_pred: Tensor,
         ce = tf.reshape(ce, [tf.shape(ce)[0], -1])
         ce = tf.reduce_mean(ce, 1)
     else:
-        y_true = y_true.to(torch.float)
+        y_true = y_true.to(y_pred.dtype)
         if from_logits:
             ce = torch.nn.BCEWithLogitsLoss(reduction="none")(input=y_pred, target=y_true.view(y_pred.size()))
         else:
             ce = torch.nn.BCELoss(reduction="none")(input=y_pred, target=y_true.view(y_pred.size()))
 
         if class_weights is not None:
-            sample_weights = torch.ones_like(y_true, dtype=torch.float)
+            sample_weights = torch.ones_like(y_true)
             for key in class_weights.keys():
                 sample_weights[y_true == key] = class_weights[key]
             ce = ce * sample_weights.reshape(ce.shape)
